+++ internal/infrastructure/db/influx/market_price_repository_impl.go
@@ -58,9 +58,9 @@ func (i *influxDbService) GetPricesForMarkets(
 		marketID := result.Record().ValueByKey(marketTag).(string)
 		marketPrice := domain.MarketPrice{
 			MarketID:   result.Record().ValueByKey(marketTag).(string),
-			BasePrice:  int(result.Record().ValueByKey(basePrice).(int64)),
+			BasePrice:  result.Record().ValueByKey(basePrice).(float32),
 			BaseAsset:  result.Record().ValueByKey(baseAsset).(string),
-			QuotePrice: int(result.Record().ValueByKey(basePrice).(int64)),
+			QuotePrice: result.Record().ValueByKey(basePrice).(float32),
 			QuoteAsset: result.Record().ValueByKey(quoteAsset).(string),
 			Time:       result.Record().Time(),
 		}
